Adding serverside authorisation for delete-quix and update-quiz

+++ src/pages/api/delete-quiz.ts
@@ -1,6 +1,8 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import { PrismaClient } from '@prisma/client'
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { getServerSession } from 'next-auth/next'
+import { authOptions } from './auth/[...nextauth]'
 
 const prisma = new PrismaClient()
 
@@ -8,42 +10,48 @@ export default async function handle(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-    const { quizid } = req.query
-    const quizidCasted = Number(quizid)
-    
-    try {
-        // Prisma docs (no date)
-        const deletedQuiz = await prisma.$transaction(async (prisma) => { // transaction ensures the deletes are executes together, not seperately
-            
-            // Order of deletion: answers -> questions -> quiz
+    const session = await getServerSession(req, res, authOptions)
+    if(session?.user.role === "ADMIN") {
+        
+        const { quizid } = req.query
+        const quizidCasted = Number(quizid)
+        
+        try {
             // Prisma docs (no date)
-            await prisma.answer.deleteMany({
-                where: {
-                    question: {
+            const deletedQuiz = await prisma.$transaction(async (prisma) => { // transaction ensures the deletes are executes together, not seperately
+                
+                // Order of deletion: answers -> questions -> quiz
+                // Prisma docs (no date)
+                await prisma.answer.deleteMany({
+                    where: {
+                        question: {
+                            quiz_id: quizidCasted
+                        }
+                    }
+                })
+                
+                await prisma.question.deleteMany({
+                    where: {
                         quiz_id: quizidCasted
                     }
-                }
+                })
+                
+                // Prisma docs (no date)
+                return prisma.quiz.delete({ // return last operation because it shows that all the previous operations were successful
+                    where: {
+                        quiz_id: quizidCasted
+                    }
+                })
             })
-            
-            await prisma.question.deleteMany({
-                where: {
-                    quiz_id: quizidCasted
-                }
-            })
-            
-            // Prisma docs (no date)
-            return prisma.quiz.delete({ // return last operation because it shows that all the previous operations were successful
-                where: {
-                    quiz_id: quizidCasted
-                }
-            })
-        })
 
-        if (deletedQuiz) {
-            return res.status(200).json({ message: 'Quiz successfully deleted' })
+            if (deletedQuiz) {
+                return res.status(200).json({ message: 'Quiz successfully deleted' })
+            }
+        } catch (error) {
+            console.log(error)
+            return res.status(500).json({ message: (error as Error).message })
         }
-    } catch (error) {
-        console.log(error)
-        return res.status(500).json({ message: (error as Error).message })
+            
     }
+
 }
